backend/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
PyMuPDF
google-generativeai
openai
anthropic
python-dotenv
pillow
firebase-admin
